Moved CrumbManager and added test
