+++ src/test/java/yahoofinance/OnlineRequestTest.java
@@ -1,0 +1,26 @@
+package yahoofinance;
+
+import org.junit.Test;
+import yahoofinance.mock.MockedServersTest;
+import yahoofinance.quotes.fx.FxQuote;
+import yahoofinance.quotes.fx.FxSymbols;
+
+import java.io.IOException;
+import java.math.BigDecimal;
+import java.util.Map;
+
+import static org.junit.Assert.*;
+
+/**
+ *
+ * @author Paul ProgramComputer
+ */
+public class OnlineRequestTest {
+
+    @Test
+    public void OnlineRequestTest() throws IOException {
+        Stock stock = YahooFinance.get("INTC");
+     assertEquals(stock.getStats().getSymbol(),"INTC");
+
+    }
+    }
